pyproject.toml
[tool.poetry]
name = "my-salary"
version = "0.1.0"
description = ""
authors = ["Dima Chubatov <dima.don84@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
psycopg2 = "^2.9.6"
databases = {extras = ["postgresql"], version = "^0.7.0"}
alembic = "^1.11.1"
email-validator = "^2.0.0.post2"
sqlalchemy-utils = "^0.41.1"
python-multipart = "^0.0.6"
pytest = "^7.3.1"
pytest-freezegun = "^0.4.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
black = "22.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
